requirements.txt
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
feedparser>=6.0.10
yfinance>=0.2.36
beautifulsoup4>=4.12.0
pytest>=8.0.0
